Fix issue 7141762: AD groups traversals can ignore interrupts when talking to LDAP servers.
Instead of treating an InterruptedNamingException in AdServer.search as
a generic NamingException, I throw it back to AdGroupsTraversalManager.run 
where it is caught and exits the main loop, ending the traversal.

At this point in the loop, we have not made any partial changes to the
database, so if the thread interruption was spurious, we will simply
resume later and pick up where we left off.

Code review: https://codereview.appspot.com/6498116/